requirements.txt
keras==2.0.8
pillow==5.4.1
pyglet==1.3.0
